Prevent existing OS_CLOUD variable messing up deployment
We're shipping a clouds.yaml and use it to connect to OpenStack.
However, we allowed setting a different OS_CLOUD than the one entry
defined in the clouds.yaml. This is not good and causes issues when you
run the script and you have OS_CLOUD already defined in your
environment variables.

This prevents issue we set the OS_CLOUD to shiftstack-bm, with the
possibility to override it with the OVERRIDE_OS_CLOUD variable.
We also don't copy the clouds.yaml to the working directory to prevent
it from taking precedence.