Add ShiftStack self-signed CA
The Certificate Authority will be used when deploying OpenStack in CI,
to sign the Certificates that will be used for the OpenStack public
endpoints.

That way, we can configure OpenShift infra to use a static CA cert, and
store the cert/key here in this repo where it's encrypted.

This patch does the following:
* activate SSL when deploying OpenStack
* store CA cert & key
* use the CA cert & key in dev-install

Notes: secrets were added to Github Actions if we deploy things via
Github CI.