Deploy 3-nodes CI with edge-style environment
Create 2 configs:
* central which will be the central site, hosting the OSP control
  plane, and compute service as well.
* az0 which will be a compute node in its own availability zone named
  az0.

They would be deployed on RHEL 8.2, OSP 16.1 using RHSM (customer
bits), with Octavia, Manila, Cinder, Ceph, SSL for public endpoints, and Availability zones.

The `provision.sh` script has been adapted to collect the data from the
central node and scp it to the az nodes at the right time.
This step is normally documented in dev-install but we automate it
here.

Finally, this commit includes secrets/osp-ci/exported-data which
contains the data that is generated on the central site, and used by
the az nodes. We store it here in case we want to deploy more AZs.