Simplify configs
We can figure out the cloudname and hostname by using the config name.
Also we don't want to pull RHCOS image automatically, but this is done
by default in dev-install so we disable it.

Finally, adding a small command about the variables specific to the
platform vs the ones in VEXXHOST.