hwoffload: add openstack-proxy for ipv6 endpoint
We are adding a new CI job to test single stack ipv6.
However, our OpenStack endpoints are only on ipv4 and it's going to be
difficult to change that in dev-install. So in the meantime, we use
`openstack-proxy` tool from
https://github.com/pierreprinetti/openstack-proxy.

This will create an endpoint on a port that is not open to the firewall,
so it's not available from the Internet. But that's fine because the
only user of this endpoint should be the OCP machines that are deployed
and are single stack v6 already and within the same subnet as the
endpoint so there is no need to open IPtables.